a bunch of tidying up the client side
