--- example/src/lib/types.ts
@@ -1,8 +1,0 @@
-export interface Message {
-  id: string;
-  chatId: string;
-  content?: string;
-  streamId?: string;
-  role: "user" | "assistant";
-  timestamp: Date;
-}
