+++ example/tsconfig.json
@@ -17,12 +17,11 @@
     "paths": {
       "@/*": ["src/*"]
     },
-
     /* This should only be used in this example. Real apps should not attempt
      * to compile TypeScript because differences between tsconfig.json files can
      * cause the code to be compiled differently.
      */
     "customConditions": ["@convex-dev/component-source"]
   },
-  "include": ["./src", "vite.config.ts"]
+  "include": ["./src", "vite.config.ts", "convex"]
 }
